Eliminó CaificarPlato.java q no hacía nada, al layout detalle plato le agregué un campo para la nota actual en la DB, a la clase plato le cambié el promedio de ser un "int" a ser un "float" y por último se agrega la funcionalidad de q cuando se cargue un plato también cargue la nota. El meter la nota en base está ahi pero no funciona.

+++ app/src/main/java/com/rejuntadosdeinge/umenu/modelo/Plato.java
@@ -10,7 +10,7 @@ public class Plato {
     private String tipo;
     private int calificaciones;
     private int total;
-    private int promedio;
+    private float promedio;
 
     public int getId() {
         return id;
@@ -61,9 +61,7 @@ public class Plato {
     public void setTotal(int total) {
         this.total = total;
     }
-    public int getPromedio() {
-        return promedio;
-    }
+    public float getPromedio() { return promedio; }
     public void setPromedio(int promedio) {
         this.promedio = promedio;
     }
